requirements.txt
crewai==0.30.0
crewai_tools==0.2.5
langchain==0.1.20
langchain_community==0.0.38
python-dotenv==1.0.1
Requests==2.31.0
sec_api==1.0.17
streamlit==1.31.0
unstructured==0.5.1
